modifications for integrations with saturn backend

+++ proxy-server/minikube-test.sh
@@ -3,6 +3,8 @@
 which docker > /dev/null || (echo "docker is not installed. Exiting." && exit 1)
 which minikube > /dev/null || (echo "minikube is not installed. Exiting." && exit 1)
 
+PROXY_AUTH=${PROXY_AUTH:-"false"}
+
 eval $(minikube docker-env)
 
 echo -e "\033[1;92mBuilding proxy image\033[0m"
@@ -12,14 +14,21 @@ echo -e "\033[1;92mBuilding test server image\033[0m"
 (cd test && docker build -t mock-auth-server:test .)
 
 kubectl apply -f minikube-test.yaml
+if $PROXY_AUTH; then
+    kubectl apply -f minikube-mock-auth.yaml
+fi
 
 echo "Waiting for all pods to be ready..."
-kubectl wait --for=condition=Ready pod/mock-auth-server
+if $PROXY_AUTH; then
+    kubectl wait --for=condition=Ready pod/mock-auth-server
+fi
 kubectl wait --for=condition=Ready pod/proxy-test
 
-kubectl port-forward svc/mock-auth-server 8888:8888 &
-auth_pid=$!
-sleep 0.1
+if $PROXY_AUTH; then
+    kubectl port-forward svc/mock-auth-server 8888:8888 &
+    auth_pid=$!
+    sleep 0.1
+fi
 
 
 
@@ -29,7 +38,9 @@ echo "
 Open this URL in your browser: $(minikube service proxy-test --url=true)/
 "
 
-read -p "Press any key to stop forwarding and tear down resources."
+read -p "Press any key to tear down resources."
 
-kill $auth_pid || true # let deletion occur when there was an issue with port-forwarding
+if $PROXY_AUTH; then
+    kill $auth_pid || true # let deletion occur when there was an issue with port-forwarding
+fi
 kubectl delete -f minikube-test.yaml
